+++ NEWS.md
@@ -1,3 +1,8 @@
+# morphemepiece.data 1.1.0
+
+* Updated with latest wiktionary data (dump from 2021-10-22).
+* Logged data creation process in data-raw.
+
 # morphemepiece.data 1.0.0
 
 * Added a `NEWS.md` file to track changes to the package.